Add GitHub Actions CI workflow
This is preferable to the Azure pipeline for a few reasons :

- Caching. This speeds up builds considerably.
- Simplicity. Actions is easier to use than Azure.
- Integration. Actions are better integrated into GitHub and don't require a separate login.
- Mac tests are now running correctly. The Mac tests on Azure have been silently failing for a long time.

My intention is to phase out the Azure setup completely once we've run this one for a while and are happy with it. Note that I've deliberately omitted some tests at this stage. They do run fine now, but will hinder progress in subsequent PRs which will build towards Python 3 compatibility and CI. We'll make sure we're running the full suite before turning off the Azure setup, but right now I've lost a ridiculous amount of time to getting the Mac builds running, and need to make progress on my primary task : Python 3.