Build : Improved version numbering scheme
For many years we've used the semantic versioning scheme as defined at http://semver.org. However, in practice, there are external forces (eg clients: artists, producers) who view the major version as a significant milestone rather than a minor ABI break. As a result we've kept the major version at 10 for nearly 3 years now (since Jun 5, 2017), to indicate that we haven't yet fully locked down compatibility on anything.

At IE we've been using Cortex 10 in production for almost 2 years and have rarely (if ever) been bitten by the occasional ABI breaks that have trickled in over that time.

This commit introduces a "milestone" version scheme equivalent to GafferHQ/gaffer. We're setting the current milestone to 10 and bumping the major version to 1, indicating this is indeed an ABI break within the Cortex 10 landscape.

We will now always increment the major version whenever introducing any backwards-incompatible change. We'll bump the milestone version whenever Cortex reaches significant milestones in terms of functionality.

It's critical going forwards that we correctly increment the right version numbers, and describe thoroughly the nature of incompatible changes in commit messages so that they can be collated into a suitable section of the subsequent release notes.

A studio could now create a $CORTEX_COMPATIBILITY_VERSION environment variable (which would initially be 10.1), and use that for the installation of dependent tools, reducing the overhead in managing development builds.

Breaking Changes
---

- Introduced a `milestone` version and decremented `IECore::majorVersion()`. Any conditional logic using the Cortex version (both compile time and runtime) should be updated to accommodate a compatibility version instead.