USDScene : Sanitise names not supported by USD
This is not foolproof, because it's still possible for two names to sanitise to the same string if all differences consist of non-alphanumeric characters. But it should be sufficient for dealing with the problems currently being faced, which all relate to purely numeric names and the use of `:` for namespacing (see https://github.com/GafferHQ/gaffer/issues/3577).

Likewise, we're not making any attempt to roundtrip back to the original names when reading the file. There is at least one [USD issue](https://github.com/PixarAnimationStudios/USD/issues/126) where they mention using metadata to store the original name so that it can be restored on read. Looking through the USD sources though, I couldn't find any evidence that this has been implemented.

It is unfortunate that universal scene description is limited to a character set sufficient only for a small minority of the population of just one planet ;)