IECoreUSD : Refactor primvar writing into PrimitiveAlgo methods
Some primitive variables have special status in USD, and are stored as specific attributes of a prim, rather than in the "primvars:" namespace as regular primvars. The original code had a special case for redirecting "P" into such an attribute, but other variables need special treatment too. Here we are setting up a scheme whereby the primitive converter loops over all variables, passing them up a chain of calls to `PrimitiveAlgo::writePrimitiveVariable()`, which picks off the special cases on the way.

This also uses `UsdGeomPrimvarsAPI` rather than the deprecated methods of `UsdGeomImageable`.