USDScene : Round trip curve type/wrap/basis
Note that the previous code read generic UsdGeomCurves objects, assuming that everything was linear. By switching it to read UsdGeomBasisCurves we gain proper support for everything Cortex can represent, but have implicitly lost support for UsdGeomNurbsCurves and UsdGeomHermiteCurves. I have assumed that a) they are rare and b) anyone who uses them will require better reading than the previous always-linear approach. We should add specific code paths for those types if necessary, but we don't currently have suitable Cortex types to map them to.

USD support an additional "pinned" wrap mode in addition to "periodic" and "nonperiodic". We have nothing to map this into in Cortex at present, but it seems worth considering in future. It is a way of requesting that phantom endpoints be added during rendering so that the curves interpolate all the way to the end.