StringAlgo : Fix interactions between `matchMultiple()` and '*'
`matchMultiple()` was implemented to avoid scanning the pattern string twice (once to find tokens and once to perform matching). Instead the tokenization was performed internally : when a match failed it would advance through to the next token and restart there from scratch. But this advance-and-restart behaviour was also being performed unnecessarily inside the recursive calls used to implement '*' matching. This caused two problems :

- Incorrect match results for calls such as `matchMultiple( "bb", "*a b" )`. Hitting '*' would eat one character from "bb" and then recurse, and the restart in the recursion would then match "b" against "b" and succeed erroneously.
- Incredibly poor performance as every level of recursion lead to a restart for every pattern remaining. This was first reported in a [post to the gaffer-dev mailing list](https://groups.google.com/g/gaffer-dev/c/XipCCWaZ2f0/m/_AHR3JHAAAAJ).

Both problems are fixed by lifting the advance-and-restart behaviour out of the main matching function and into `matchMultiple()` itself so that it doesn't occur in the recursive calls. The new code still scans the string only once by having `matchInternal()` update the pattern pointer in preparation for `matchMultiple()` to advance it to the next token. I do wonder if this is necessary though - it's tempting to refactor `match()` to take a `string_view()` so that `matchMultiple()` can just call that directly after performing its own scan for tokens.