USDScene : Don't set default prim when writing tags
I don't know why we were doing this. The default prim is the one that is brought in by a reference by default, and that has nothing to do with tags. I can only guess that at the time this code was written, it seemed convenient that it could be retrieved again by `GetDefaultPrim()`. We now just retrieve it by name instead.