Image Tests : Replace png test file
The previous file seems to have come from kodak and was added to Cortex in 2011. Its current purpose is simply to verify that we can load png files in Nuke. The fact that it is quite large happened to expose a non-Nuke related bug in OIIO 2.0, but we now have another test verifing that bug, making this file rather large for its intended purpose.

I've swapped out the png for a smaller one, which was created using Nuke 12.1 by reading `test/IECoreImage/data/jpg/uvMap.512x256.jpg` and writing it to png with default settings on the Write node.

I've also added a similar test case to IECoreImage.ImageReaderTest, so that we may hit future issues outside of DCC specific builds.