USDScene : Optimise SdfPath -> SceneInterface::Path conversion
Rather than format a string only to split it again, we can instead pluck off the path elements one by one. This knocks ~40% off the time taken to read a large set. It should also reduce memory usage, as `SdfPath.GetString()` is documented as making a persistent string that will never be freed.