USDScene : Remove unnecessary polymorphism
I think the Reader/Writer split was probably based on IECoreAlembic, where we need a split of some sort because of the separation of the Alembic `I*` and `O*` classes for reading and writing respectively. But USD allows us to both read and write via the API objects, so there is no need for the split, as demonstrated by the fact that the `root()` and `getTime()` overloads had identical implementations.