USDScene : Rewrite tags implementation
This writes tags as UsdCollections in the main prim hierarchy, not on a sidecar "cortexTags" prim. That means that tags can now be loaded via the sets API, providing a path for transitioning away from tags and onto sets in the future. They can also now be properly referenced using the standard USD mechanisms.

It also buffers tags as PathMatchers for both reading and writing, giving much improved performance. Write time is reduced by 90% for tags at ~20000 locations, and read time is reduced by more than 99% for tags at ~4000 locations (I didn't have the patience to do a longer read test with the original code).

There is some memory overhead from the buffering, but since SceneCache stores tags in the index which is loaded on file open, it shouldn't be much different than `.scc`. I think this is fine given that we want to transition to sets anyway.