USDScene : Remove arbitrary attribute support
This is a case of taking one step back before we can take a step forwards. The implementation being removed worked by squirreling all attributes into a "cortex:" namespace. This did have the benefit of unambigous round-tripping, but at the expense of any interaction with the broader USD ecosystem. Rather than a private namespace, we need to be able to read and write attributes that might not be namespaced at all, or might be written by add-on schemas like UsdRiStatementsAPI or the schemas in the USD Arnold project. Here we're just giving ourselves a clean slate to figure out what that might involve.