Houdini Cob Reader : Use ObjectReader explicitly.
This avoids a crash in Houdini 18 when using OIIO 2.2, where for whatever reason it was attempting to read a corrupt cob file using OIIO's DPXInput and segfaulting.

We're dropping support for pdc and ptc here, but I'm not aware of anyone using those formats with this tool in recent years.