USDScene : Warn for collection members not below the collection's prim
The SceneInterface `readSet()` and `writeSet()` methods deal exclusively in paths that are relative to the current location. But USD allows a collection to reference paths that are not parented beneath the prim holding the collection. We now detect this latter case in `readSet()` and emit a warning rather than adding the wrong path to the returned set.

It's yet to be seen whether this kind of collection is prevalent in the wild, and whether or not we will need to support them fully. This [usd-interest discussion](https://groups.google.com/u/2/g/usd-interest/c/O-cKw3yr-YY/m/MOiqWWbxAQAJ) implies that we are at least not alone, in that Katana has similar needs. If we do need to support these non-self-contained collections in the future, we have a couple of options :

- Omit the non-relative paths when reading the local set, but include them when calling `readSet( includeDescendantSets = true )` for a parent location. This would keep the relative-paths API we currently have.
- Change `readSet()` and `writeSet()` to deal only in absolute paths.

It seems that it would be premature to make a call on that just yet though.