IECoreUSD::PrimitiveAlgo : Apply UsdSkel linear blend skinning
We detect prims with UsdSkel bindings and apply linear blend skinning when loading.

Note we are not exposing UsdSkel via the Cortex API, we're just baking the deformation onto the Cortex Primitive.

The arm example in the test was copied from the section "Skinning an Arm" on https://graphics.pixar.com/usd/docs/api/_usd_skel__schema_overview.html and modified to add the more modern SkinBindingAPI metadata (which avoids deprecation warnings but doesn't otherwise change the results as of USD 20.11).