IECoreUSD::PrimitiveAlgo : Adjust skinned points with bind transform
This change fixes several shapes in the Pixar HumanFemale example. Most shapes in that example have identity geomBindTransforms and are unaffected. However, the shoes, socks, and shirt buttons all have non-identity geomBindTransforms. Prior to this commmit they were floating off of the character (and the shoes were shrunken as their geomBindTransform also includes a downscale). With this change all of those shapes now stick to the character and match the visual result of usdview.

Note this transformation differs from UsdSkelBakeSkinning example code as documented via inline comment. We may need to revisit this if/when example data demonstrates a problem.