GitHub CI : Test IECoreArnold, IECoreAppleseed and IECoreVDB
Except on Mac, where we have problems. The main problem is a refusal to
find/load `libai.dylib` when running the Arnold tests. This runs OK
locally but fails on the GitHub workers, with my best guess being that
it is a SIP thing. Tried various things, including using `install_name_tool` to
remove the `@rpath` from `libai.dylib`, but to no avail.