update IE config options to support RV 2021.x.x
Some background on why we were overriding boost paths and suffix:
- RV can use Cortex built with a different version of boost.
- RV cannot use Cortex built with the same version of boost, but a
  different suffix. It causes "double free" crashes on exit.

RV versions < 7.8.0 used a different version of boost than Cortex's
default one at the time, so they didn't have problems.

RV version 7.8.x and 7.9.x use the same boost version as the current
default, centrally installed version, but with a different (empty)
suffix. So we have to use the the one provided by RV.

RV 2021.0.x uses 1.70.0, which is not compatible with Cortex at the
moment, so we have to go back to use the centrally installed one.

Considering that we are sometimes forced to use our own libraries and
sometimes the ones provided by RV, these changes should give us more
flexibility in managing incompatibilities by setting or not setting
`boostVersion` and `OpenImageIOVersion` in the IE Registry for each
given RV version, rather than having to update the options file every
time.
Until the VFX Platform solves this problem for us, of course.