SceneCacheDataAlgo: Added utility to transform path/name to and from their internal form.
We are using an internal path that append an additional root to the hierarchy
to work around limitation of USD.

Reference omit the primpath that is referenced in the resulting hierarchy.
Reference without an explicit prim path is using the default prim
and thus only can reference a single root child.

Some characters are invalid inside USD ( hyphen and parenthesis ), we need
to substitute them for internal name of collection or hierarchy's child.
Since only underscores are valid, we round trip by converting illegal character
to a number of underscore.

Note that some scenario are not supported like `-(` vs `(-` or simply
any illegal characters prior or following any number of underscore ( `-_` or `_-`).