add library paths to IECoreImage test env
`$OIIO_LIB_PATH` was not specifically added to IECoreImage test env, so
in some cases the tests would not find libOpenImageIO. That would happen
for example if `$OIIO_LIB_PATH` didn't match any of the paths that are
already specified for the general test env, like `$BOOST_LIB_PATH`, etc.

Geting the lib paths from `imageEnvPrepends` should keep the build and
test environments in sync and ensure the tests can find the same
libraries that were used in the build.