FromHoudiniPointsConverter : Revert behaviour from ImageEngine/cortex#1157
We continue to export empty PrimitiveVariables when there are no points, because downstream processing often requires the existance of P at a minimum.

An alternative to consider would be to register a FromHoudiniNullConverter which outputs an `IECore.NullObject` instead of a `PointsPrimitive`. I'm unsure of the current implications of this though, so I've left it as a todo instead.