IECoreScene::PrimitiveVariable::IndexedView : Add explicit conversion to bool and isValid() function
The current api assumes that the data pointer of an IECoreScene::PrimitiveVariable::IndexedView object
is always valid however a default constructed object has a null data pointer and calling any function
on such an object results in dereferencing a null pointer. This commit adds explicit conversion to bool
and an isValid() function that can be called to determine whether a given object is valid.

As an aside this change means that IECoreScene::Primitive::variableIndexedView() could simply return an
IECoreScene::PrimitiveVariable::IndexedView object directly instead of wrapping in a boost::optional.