ExceptionBinding : Fix exception reference counting
This lead to exceptions never being destroyed, which is particularly bad in Python 3, where every exception has a `__traceback__` attribute which holds the entire stack, keeping all objects in that stack alive forever. This was the cause of the bug exposed in https://github.com/GafferHQ/gaffer/pull/4282, where a StringPlugValueWidget isn't destroyed appropriately if it has seen a ProcessException as a result of a call to `StringPlug.getValue()`.

The Python docs are somewhat ambiguous about the reference counting requirements of `PyErr_SetObject()`, but examination of the source code shows that the reference count is incremented internally, so we do not need to increment it externally :

Python 3 :

https://github.com/python/cpython/blob/3.7/Python/errors.c#L97

Python 2 :

https://github.com/python/cpython/blob/c2f86d86e6c8f5fd1ef602128b537a48f3f5c063/Python/errors.c#L56