Alembic PrimitiveWriter : Fix indices handling in `geomParamSample()`
The `UInt32ArraySample` references the sample data rather than copying it, so we can't give it a temporary that goes out of scope before the sample is written. Fortunately the layout of positive integers is the same as unsigned integers, so we can use a cast to reference the original data instead of a temporary.