IgnoredExceptions : Clear `__traceback__`
In Python 3, the new `__traceback__` attribute on exceptions means that the exception contains a circular reference to itself (because the traceback references the stack frame where the exception exists as a local variable). And worse, the traceback keeps all the calling stack frames and all their local variables alive too.

This was causing problems in Gaffer, by keeping `GafferUI.Widgets` alive well after the underlying QWidget was destroyed, leading to the dreaded `C++ object already deleted` error.

Since the purpose of IgnoredExceptions is to hide the exception, it is fine erase the traceback (since nobody will see it anyway).