ToGLStateConverter : Convert `Cs` attribute to `Color` StateComponent
This will then be picked up by the `Cs` parameter value for any GLSL shaders, unless overridden by a `Cs` primitive variable. We should have been doing this all along, but it is crucial to displaying colours in the viewport now that IECoreUSD is loading constant primvars as attributes instead of primitive variables.