FromHoudiniCurvesConverter: Updated to test to follow new expectation for GU_Detail with a mix of open/close poly.
Following our next expectation, when a GU_Detail is a mixture of open and closed poly
with multiple locations, we want to convert the open poly as CurvesPrimitive and
the closed one as MeshPrimitive.

We believe the previous behaviour was to maintain backward compatibility when
we introduce support for converting a GU_Detail made only of open poly
as a CurvesPrimitive.

We don't think it makes anymore to convert a mixture of open/close poly
as mesh as it's part of a workflow to export say a tree mesh along
side with the curve used to drive it's dynamic motion.