CompoundData : Avoid duplicated call to readable()
I'm pretty sure we've seen cost associated with readable() before?  I don't think it would be particularly significant, but there doesn't seem to be any harm in removing a call.  This sort of commonly used wrapper class wouldn't be obvious in a profile if it was contributing a little inefficiency.