CI : Update to GafferHQ/dependencies 5.0.0a3
This is what we are using for the Gaffer 0.62 test builds.

The MacOS `ImageIO.framework` workaround in the SConstruct mirrors a workaround we have had in the Gaffer wrapper for a long time. I don't know why it has only become necessary for the Cortex tests now. We also need to reset `OIIO_LIB_PATH` in `options.posix` to avoid the dependency libraries from getting re-inserted in front of `ImageIO.framework` in `imageTestEnv`. This could be avoided by refactoring the logic for building `imageTestEnv`, but I'm unsure what knock-on effects that might have for an Image Engine build. For now I'm trying to limit the changes to MacOS/GafferHQ.