USDScene : Support `doubleSided` attribute
Write support is not ideal, because USD will only author `doubleSided` onto Gprims. So it is impossible to write onto transform-only locations, and `writeObject()` must be called before `writeAttribute()` on geometric locations. It may be possible to improve on this by buffering attributes and automatically inheriting them from transforms onto primitives, but it's not clear that the complexity is justified at this point.