SceneShapeProxy: Add a proxy for `SceneShape`
We noticed huge FPS drops in scenes with a huge amount of
`SceneShapes`. Profiling showed that a lot of time is spend by
`SubSceneOverride` requesting draw updates from the shapes in
the scene. This also happens for shapes that are set as hidden,
what we consider a bug/flaw in Maya's VP2 drawing API.

To alleviate this issue, we derive the proxy from `SceneShape`
and inherit all of it's behaviour with the exception, that we don't
register it as drawable. This allows us to replace `SceneShape`s with
its proxy implementation, where only the data reading capabilities
are needed, e.g. layouts or rigs.