IECoreMaya : Fixed erroneous display of hidden locations in VP2.
The VP2 scene traversal is depth first, and we were checking for visibility
overrides after the child traversal. So the child locations which have
scene:visible set to true, were being displayed and the location which had
an override was being hidden.

Now the check for a visibility override comes before the traversal, and the
child traversal is pruned if the location is invisible.