IECoreMaya : Fixed SceneShape isolate selection in maya.
This commit fixes the isolate selection of SceneShape objects. An even
better improvement would be to support isolate selection of components
within a collapsed SceneShape. From an initial investigation, I
believe the issue first stems from not being able to add SceneShape
components to a maya object set. The documentation is not entirely
clear on what is involved in this process, but the "Geometry Data"
subsection of the following article may hold some clues.

https://around-the-corner.typepad.com/adn/2012/09/custom-shapes-in-the-maya-api-1.html

The crux of the situation may depend on implementing an MPxGeometryData
for SceneShape (which is a bit crude since the scene interface may
contain an entire scene, not just a single shape). However the blog
seems to indicate that "Set information for components is stored
with the geometryData".