IECoreMaya : Updated SceneShape drawing to respect ancestral visibilty.
There have been visibility discrepancies when a layout is collasped
or expanded. When collasped, the drawing code is able to read
visibility overrides and hide the appropriate sub-hierarchies.
However, when the same layout is expanded the visibility was tested
at each expanded location, and it failed to consider overrides
which occured further up in the hierarchy.

This commit solves this discrepancy by using LiveScene to determine
if the sceneShape is should actually be invisible due to an
override further up the hierarchy. SceneShape registers a custom
readAttribute function with LiveScene, so LiveScene is able to
correctly read the visibility overrides during its traversal.

In regards to performance, when a SceneShape is fully expanded, and
there is a deep hierarchy of visible transforms, this commit will
slightly impact the frame rate since LiveScene will be performing extra
work to ensure there are no visiblity overrides. Some of this overhead
may be reduced through the use of an extra visiblity tracking cache,
but there is also a price to pay for the extra message callbacks
necessary to keep the cache synchronized. On a fully visible robot with
~2870 locations, the frame rate dropped ~1.2 fps.

However, in practice at IE, this will not affect the performance of our
animation rigs since all SceneShapes will be expanded to maya geometry.
In this scenario, MDagPath::isVisible() will return False since
the SceneShape will be set to an intermediate object, and the LiveScene
code path will not be executed.