AlembicScene : Support various flavours of integer GeomParams
I've omitted CharVectorData/ICharGeomParam as we are unable to support them using the generic code we currently use. CharVectorData is templated on `char` and ICharGeomParam is templated on `unsigned char` and these are distinct types. Really we should define our TypedData types using `[u]int_[8|16|32]` rather than `char|short|int`, but that is a job for another day.