SceneShapeTest: Create maya node in `setUp` method
Moving the node creation in the `setUp` method, which is called before
every test method, gives subclasses the ability to run the same test
cases with a different node. This is e.g. useful for
`ieSceneShapeProxy` as we need to make sure that it behaves the same
way as it's non proxy version