SceneCacheReader: Stop transforming the mesh into world space.
We want to maintain the world space matrix instead of transforming the points
so we can more easily round trip scene cache in/out of nuke.
This allows to write the bounds as the SceneInterface expects ( local space ).

Due to some kind of reset happening in the SourceGeo source code, we need to store
the matrix in create_geometry and apply in the geometry_engine.

I think this should work fine considering that Op are instantiate per output context ( frame )
so we shouldn't have a clash in the map data structure.