SceneShapeSubSceneOverride: Check visibility for root locations
In case we have a root location i.e. we set a shape node to
`objectOnly` and we also set "draw root bounds", the visibility
of the current instance was never taken into accout and we draw the
bounds regardless. We address this and skip adding a render item for
every instance that is set to invisible.