LiveScene: Added caching for GeometryList.
Querying Nuke repeatedly for a GeoOp GeometryList proved to be really slow
for large hierarchy.
We introduce a caching mechanism using nested map that cache at 3 levels:

- LiveScene*
- Op hash
- frame

LiveScene allows to maintain a global cache between different LiveScene and avoid collision
between the LiveScenes' input GeoOp.
Op Hash allows to cache for the same Op but with different knob values or connection upstream.
Finally, frame allows to cache all of the above for a given frame.

When trying to get the GeometryList, we first check if we have any cache hit for the current LiveScene,
then we try to find a cache for the current Op hash and finally for a specific frame.
If we have a cache miss, we call the actual geometryList( frame ) to get the input GeoOp actual GeometryList
and then we add that to the cache before returning the cached GeometryList.

To make sure the caching is fast enough, we are avoid hashing using something like MurMurHash and we also
want to maintain that nested structure because of the pattern for cache invalidation.
For example, if an Op Hash is modified ( input connection changed upstream ), it's likely we can throw away all the cached frames for it.