MeshAlgo::triangulate : Refactor and fix reserve() sizes
The refactor isn't needed - I had originally thought it might make sense to expose a function that just returned the indices needed by distributePoints, without the overhead of actually reindexing the data - before I realized that there's actually no need for triangulate to reindex the data at all.

I can stuff this function back into where it came from if you prefer it, but I find it a little clearer with a separate function that handles the indices. The only difference in behaviour is that it now sets the reserve sizes correctly ( it would be easy enough to do this to the original code if we do revert the refactor ).