TriangleAlgo::triangleClosestBarycentric : Fix numerical precision bug
This code claims to be copied from www.geometrictools.com, however, somewhere in the coures of it being copied ( or perhaps in an earlier version of geometrictools ), someone got a little overly clever with optimization.

The version currently on geometrictools.com uses 26 different cases ( depending on which part of the triangle is closest ) to compute the barycentric coordinates, and then uses the barycentric coordinates to compute the closest point, and does a straightforward distance calculation from the closest point.

The previous version of this function instead used a different way of computing distance for each of the 26 cases, sometimes it is able to reuse calculations that have been done to figure out which case we're in.  This does save a few floating point operations ... however it suffers from very catastrophic cancellation. In some cases, it is taking the squared distance from an arbitrary vertex, and then subtracting a squared distance representing the portion of the distance which is parallel to the target feature. If the actual closest distance is small, the squared closest distance is very small, and computing this by subtracting two large numbers is highly inaccurate.

This results in MeshPrimitiveEvaluator::closestPointWalk choosing the wrong polygon, and telling me that the points I'm generating on a unit cube are 0.0005 away from the surface, and me thinking there has to be a bug in my code.

The solution is to switch to the approach currently used by geometrictools: use the barycentric coordinates to get the closest point, then directly compute the distance from that. This is hypothetically a few extra operations, but only adds and multiplies ... and it saves us a call to fabs because we now compute the squared distance using a method that cannot be negative. The loss in performance is probably too small to measurable, but even if there is a small loss in performance, the increase in accuracy ( and simplicity of the code in not having 26 version of the distance calculation ) would be worth it.