MeshAlgo::connectedVertices : Optimize
This is irrelevant to the task at hand, but I played with it on the weekend - mostly because it lent itself to really theoretically elegant approach: because the storage we need for the maximum possible number of adjacent vertices based on the number of polygons touching a vertex is twice the average number of vertices in practice, and because the id of adjacent vertices are likely to differ in their low order bits, this should be a very efficient place to use a very simple hash table based on a modulo of the vertex id.

However, in practice, that proved 50% slower on average meshes than just using linear insertion. I would have expected linear insertion to completely break down on vertices with very high number of neighbours ... but in about the most extreme case I can think of ( a sphere with just 10 rings of latitude, but 10000 vertices connected to the poles ), linear insertion is only 10X slower. I think that means that there really isn't any point in using the hash table.

Compared to the previous implementation with an std::set per vertex, this is twice as slow on the extreme vertex test case, but 4X faster on reasonable meshes.