SceneShape: Fix reading of time samples for expanded link locations
When reading a link at a location the determination if time is added
to the `linkAttributeData` was solely based on the connection of the
time plug. In case the time plug of the scene shape is not connected to
maya's global time output, we assumed a time remapping to happen.

This leads to adding the current time to the `linkAttributeData` if
the location is an expanded child location, because we connect the
out time of a scene shape to it's childs in time on expansion, which
breaks this assumption.

We address this issue by traversing the time plug's connection chain
and check if the plug is connected to any scene shape or maya's global time