FrameRange : Prevent creation of FrameRanges with negative steps
Previously, you could create a FrameRange object with a negative step,
but it would go into an infinite loop if you tried to get the frame list

Given that we have the ReversedFrameList that achieves the same as a
negative step, it felt like using that was the correct approach.

Note that we still include the support for the negative step in the
parse method, so that we can provide a more useful error message during
init.