DataTraits.dataTypeFromElement : Convert lists to preferred data type
When converting a list, it was simply using the first match, which would
often not be the desired one. Ex: converting `["abc", "b"]` to
`IECore.CharVectorData( [ 97, 98 ] )`, instead of
`IECore.StringVectorData( [ 'abc', 'b' ] )`.

The solution here is to use, on lists, the already existing flag on the
`__dataTypesConversionDict` entries that indicated the preferred
IECore.Data type for a given source.

The options selected here as the preferred one when there was ambiguity
were the same ones as returned by the simple (non-vector) types.

That required a small update to `dataTypeFromElementType()` in order to
avoid returning a random return, now that `list` entries in
`__dataTypesConversionDict` can have a preferred flag set to `True`.